refactor(tools): modularize registry module
